wrote spd emulator tests
I have the tests all commented out and I didn't add the spd emulator as a project reference because doing so would require updating reloaded.memory to the newest version.

The problem with this is it means every dependency needs to be updated, and the dependencies for the dependencies need to be updated, and code needs to be changed because of changed/removed methods. I spent about an hour trying to do it and gave up :/